+++ app/client/src/utils/autoHeight/generateTree.ts
@@ -1,6 +1,7 @@
 import { areIntersecting } from "utils/boxHelpers";
 import { pushToArray } from "utils/helpers";
 import { MAX_BOX_SIZE, NodeSpace, TreeNode } from "./constants";
+import { getNearestAbove } from "./helpers";
 
 // This function uses the spaces occupied by sibling boxes and provides us with
 // a data structure which defines the relative vertical positioning of the boxes
@@ -76,6 +77,7 @@ export function generateTree(
       if (originalBottomRow === undefined || layoutUpdated) {
         originalBottomRow = currentSpace.bottom - MAX_BOX_SIZE;
       }
+
       tree[currentSpace.id] = {
         aboves: aboveMap[currentSpace.id] || [],
         belows: belowMap[currentSpace.id] || [],
@@ -83,7 +85,16 @@ export function generateTree(
         bottomRow: currentSpace.bottom - MAX_BOX_SIZE,
         originalTopRow,
         originalBottomRow,
+        distanceToNearestAbove: 0,
       };
+    }
+  }
+
+  for (const boxId in tree) {
+    const nearestAbove = getNearestAbove(tree, boxId, {});
+    if (nearestAbove.length > 0) {
+      tree[boxId].distanceToNearestAbove =
+        tree[boxId].topRow - tree[nearestAbove[0]].bottomRow;
     }
   }
